+++ moban/utils.py
@@ -3,9 +3,7 @@ import re
 import sys
 import stat
 import errno
-
 import yaml
-
 import moban.constants as constants
 import moban.exceptions as exceptions
 
@@ -147,3 +145,19 @@ def pip_install(packages):
     subprocess.check_call(
         [sys.executable, "-m", "pip", "install", " ".join(packages)]
     )
+
+
+def get_template_path(template_dirs, template):
+    temp_dir = ''
+    for a_dir in template_dirs:
+        if os.path.exists(
+                os.path.join(a_dir, template.filename)) and \
+                os.path.isfile(
+                    os.path.join(a_dir, template.filename)
+                ):
+            temp_dir = a_dir
+            break
+    temp_file_path = os.path.join(
+        os.getcwd(), os.path.join(temp_dir, template.filename)
+    )
+    return temp_file_path
